fix: tx scopes so that execute now is visble directly
