+++ dhis-2/dhis-services/dhis-service-core/src/main/java/org/hisp/dhis/scheduling/JobScheduler.java
@@ -179,7 +179,7 @@ public class JobScheduler implements Runnable, JobRunner {
       String jobId = jobIds.poll();
       while (jobId != null) {
         JobConfiguration config = service.getJobConfiguration(jobId);
-        if (config != null && config.getJobStatus() == JobStatus.SCHEDULED) {
+        if (config != null && (config.getJobStatus() == JobStatus.SCHEDULED)) {
           Instant now = Instant.now().truncatedTo(ChronoUnit.SECONDS);
           Instant dueTime = dueTime(now, config);
           runDueJob(config, dueTime);
